% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kalmanST.R
\docType{package}
\name{kalmanST-package}
\alias{kalmanST}
\alias{kalmanST-package}
\title{MedQuestions}
\format{\preformatted{ logi NA
}}
\usage{
kalmanST
}
\description{
Blah
}
\details{
\tabular{ll}{
Package: \tab kalmanST\cr
Version: \tab 0.1\cr
Date: \tab 2016-01-04\cr
License: \tab GPL-3 \cr
}



section{Dependencies}{
The package \code{kalmanST} depends upon some other important contributions to CRAN in order to operate; their uses here are indicated:\cr\cr
    blah
}

section{Citation}{
blah
}

references{
X
}
}
\author{
Benjamin M. Taylor, Department of Medicine, Lancaster University
}
\keyword{package}

