% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kfadvance.R
\name{kfadvance}
\alias{kfadvance}
\title{kfadvance function}
\usage{
kfadvance(obs, oldmean, oldvar, A, B, C, D, E, F, W, V, marglik = FALSE,
  log = TRUE, na.rm = FALSE)
}
\arguments{
\item{obs}{X}

\item{oldmean}{X}

\item{oldvar}{X}

\item{A}{X}

\item{B}{X}

\item{C}{X}

\item{D}{X}

\item{E}{X}

\item{F}{X}

\item{W}{X}

\item{V}{X}

\item{marglik}{X}

\item{log}{X}

\item{na.rm}{X}
}
\value{
...
}
\description{
A function to
}

