% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kfadvance1.R
\name{kfadvance_stationary_STGP}
\alias{kfadvance_stationary_STGP}
\title{kfadvance_stationary_STGP function}
\usage{
kfadvance_stationary_STGP(obs, oldmean, oldvar, A, B, C, D, E, F, W, V,
  marglik = FALSE, log = TRUE, na.rm = FALSE)
}
\arguments{
\item{obs}{X}

\item{oldmean}{X}

\item{oldvar}{X}

\item{A}{X}

\item{B}{X}

\item{C}{X}

\item{D}{X}

\item{E}{X}

\item{F}{X}

\item{W}{X}

\item{V}{X}

\item{marglik}{X}

\item{log}{X}

\item{na.rm}{X}
}
\value{
...
}
\description{
A function to
}

