% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{kffit_stationary_STGP}
\alias{kffit_stationary_STGP}
\title{kffit_stationary_STGP function}
\usage{
kffit_stationary_STGP(form, param, pid, tid, tmax, data, cov.model,
  history.means = TRUE, history.vars = TRUE, diagonal_only = TRUE,
  prior.mean = NULL, prior.var = NULL, fit = TRUE, se.fit = TRUE,
  se.predict = TRUE, noisy = TRUE, na.rm = TRUE)
}
\arguments{
\item{form}{X}

\item{param}{list of parameters, see ?paramSpec}

\item{pid}{X}

\item{tid}{X}

\item{tmax}{X}

\item{data}{X}

\item{cov.model}{X}

\item{history.means}{X}

\item{history.vars}{X}

\item{diagonal_only}{X}

\item{prior.mean}{X}

\item{prior.var}{X}

\item{fit}{X}

\item{se.fit}{X}

\item{se.predict}{X}

\item{noisy}{X}

\item{na.rm}{X}
}
\value{
...
}
\description{
A function to
}

