% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{KFparest}
\alias{KFparest}
\title{parest function}
\usage{
KFparest(fun, ...)
}
\arguments{
\item{fun}{X}

\item{...}{optinal arguments to be passed to optim}
}
\value{
...
}
\description{
A function to
}

