% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parameterEstimation.R
\name{paramSpec}
\alias{paramSpec}
\title{paramSpec function}
\usage{
paramSpec(a = NULL, sigma_S = NULL, phi_S = NULL, sigma_Y = NULL)
}
\arguments{
\item{a}{X}

\item{sigma_S}{X}

\item{phi_S}{X}

\item{sigma_Y}{X}
}
\value{
...
}
\description{
A function to
}

