% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/postprocess.R
\name{predictPlots}
\alias{predictPlots}
\title{predictPlots function}
\usage{
predictPlots(out, pause = TRUE, ...)
}
\arguments{
\item{out}{X}

\item{pause}{X}

\item{...}{X}
}
\value{
...
}
\description{
A function to
}

