% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{run_filter}
\alias{run_filter}
\title{run_filter function}
\usage{
run_filter(kf_spec, optim = FALSE, history.means = FALSE,
  history.vars = FALSE, diagonal_only = TRUE, prior.mean = NULL,
  prior.var = NULL, fit = FALSE, se.fit = FALSE, se.predict = FALSE,
  noisy = TRUE, na.rm = FALSE)
}
\arguments{
\item{kf_spec}{X}

\item{optim}{X}

\item{history.means}{X}

\item{history.vars}{X}

\item{diagonal_only}{X}

\item{prior.mean}{X}

\item{prior.var}{X}

\item{fit}{X}

\item{se.fit}{X}

\item{se.predict}{X}

\item{noisy}{X}

\item{na.rm}{X}
}
\value{
...
}
\description{
A function to
}

